requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# 数据库
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# 数据处理
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# 可视化
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# 机器学习
scikit-learn==1.3.2

# Redis
redis==5.0.1
hiredis==2.2.3

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1
ruamel.yaml==0.18.5

# 加密
cryptography==41.0.8

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 日志
loguru==0.7.2

# 时间处理
python-dateutil==2.8.2

# 文件处理
openpyxl==3.1.2
xlsxwriter==3.1.9

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 监控
prometheus-client==0.19.0

# 任务队列
celery==5.3.4

# 其他工具
tabulate==0.9.0
tqdm==4.66.1
